+++ src/client/asynchronous.rs
@@ -926,6 +926,7 @@ mod tests {
             #[cfg(feature = "embedded-core")]
             embedded_server: None,
             response_timeout: None,
+            connect_timeout: None,
         };
         assert!(FalkorAsyncClient::build_executor(
             &mut provider,
@@ -962,6 +963,7 @@ mod tests {
             #[cfg(feature = "embedded-core")]
             embedded_server: None,
             response_timeout: None,
+            connect_timeout: None,
         };
         assert!(provider.has_sentinel_replica());
         assert!(FalkorAsyncClient::build_executor(
@@ -989,6 +991,7 @@ mod tests {
             #[cfg(feature = "embedded-core")]
             embedded_server: None,
             response_timeout: None,
+            connect_timeout: None,
         };
         let readonly_conn = provider
             .get_async_connection()